requirements.txt
pycrypto==2.6.1
requests==2.9.1
simple-crypt==4.1.7
wheezy.caching==0.1.114
wheezy.core==0.1.140
wheezy.html==0.1.147
wheezy.http==0.1.344
wheezy.routing==0.1.157
wheezy.security==0.1.64
wheezy.validation==0.1.135
wheezy.web==0.1.485
